refactor : compareCreatedTimeAndNow() 메소드를 TimeUtil 클래스로 분리
